requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==4.1.0
boto3==1.28.15
botocore==1.31.15
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==4.2.0
django-database-url==1.0.3
django-debug-toolbar==4.1.0
django-filter==23.2
django-redis==5.3.0
django-silk==5.0.3
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
flower==2.0.0
gprof2dot==2022.7.29
gunicorn==20.1.0
humanize==4.7.0
idna==3.4
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.1
kombu==5.3.1
MarkupSafe==2.0.1
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.13.3